Add files via upload
## Assignment-3-4-Critique-by-Design

#so I am interested in healthcare, so I searched the CDC’s website for some graphs I could improve upon. I found some on health insurance coverage, race, and year that I thought could be redone to be more visually enticing and tell a better story. 

https://www.cdc.gov/nchs/ppt/hus/hus18fig20.pptx

https://www.cdc.gov/nchs/data/hus/2018/fig20.pdf

https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/Health_US/hus18figures/fig20.xlsx

It was an interrelated group of 2 graphs, so I decded to do both, although, in hindsight, I probably should’ve just stuck to the second, which is where I saw more opportunity. The first chart was a chart on how insurance coverage has changed from 2007 to 2018. The second was a graph of the different types of insurance coverage people of different races had in 2017. The data on the first chart was just presented in a line graph, which I didn’t think was very clear or compelling, and the second graph didn’t effectively show the disparities between races, so I took to recreating them both. 

I use powerpoint to create almost all my graphs for several reasons. It is able to create clear, nice looking graphs quite easily; it integrates well with excel, which is good for simple data analysis, which is what I do most of the time; it can auto-create graphs based on the data fed into it (though they always need to be fine-tuned), the graphs can be edited by a lay person with no coding skills, which is great when you want someone else to work on your graphs as well, and almost everyone I know has the Microsoft office suite on their computer. 

The first iteration of the first chart was a stacked area chart. This proved to be confusing for both my testers, as they saw the lower areas  as having less than the higher areas, because they pictured it as a sort of series of half full pictures layered on top of one another. So using this feedback, I switched to a bar chart and labeled the percentages on each bar. It’s not as neat as the stacked area chart, but it does look better than the original, and it is clearer as well, at least from my users input. I had immediate understanding for the second chart, so I pretty much left it as is.